Do not set subpreaper on exec processes
Because an exec'd process needs to be reparented to the pid 1 of the
container the subreaper will cause issues with this and cause the parent
to stay outside of the namespace.

This makes runc on exec long running but there is not a better way to do
with when we do not have the subreaper reparent.

Signed-off-by: Michael Crosby <crosbymichael@gmail.com>