Write rusage to a json file on process exit
The rusage data is already available but dropped when waiting for the
process to exit. This change writes the rusage to the containerd state
dir with the <container id>-stats.json filename.

Caveat: it is now possible to fill the work directory of containerd
(usually in /run, a tmpfs) by spawning millions/billions of containers,
since there is no cleanup mechanism involved. A potential solution would
be to write those somewhere with more room for them, or having a
periodical cleanup.

Signed-off-by: Mathieu Pasquet <mathieu.pasquet@alterway.fr>