fix event lost after continuous containerd restart
kill container shim when containerd is down, once containerd
started, it will generate a new exit event and write it to event.log.
docker won't get this event if the event connection is not established.

containerd got killed and started again, it will read events from
event.log and then eliminate all the data in it with trunc flag.
By now, if docker is still unconnected with containerd, it cannot get
this event either.

containerd get killed and started for the third time. as event.log is
empty, containerd can not get the exit event. even connection between
docker and containerd is established successfully, event is lost and
never passed to docker daemon.

To repro this, needs:
 * kill shim when containerd is down, to generate exit event
 * connection between docker and containerd is down for a long time
 * containerd got killed and restartd for more than 3 times

this patch remove trunc flag to avoid this.

use event.log.tmp to avoid event missing

Signed-off-by: Deng Guangxing <dengguangxing@huawei.com>