Add transaction to containerd
Why we need transaction in containerd:
   When docker send grpc messages to containerd, it always forward to exec runc to do something.
   And then send back a event to docker.

But if the containerd was killed or crashed after receving grpc message, before sending back the events.
These is no way to recover it. This may cause the data of container not the same between docker and containerd.

How to implement:
 We add a transaction mechanism:

* Open a transaction:
This will return a Transaction interface
```
 transaction,err := transactionFactory.OpenTransaction(containerID, transactionType, otherMetada)
```

* Close the transaction:
```
 transaction.Close()
```

Signed-off-by: Wentao Zhang <zhangwentao234@huawei.com>