metadata: define content sharing policy
This changeset modifies the metadata store to allow one to set a "content
sharing policy" that defines how blobs are shared between namespaces in the
content store.

The default mode "shared" will make blobs available in all namespaces once it
is pulled into any namespace.  The blob will be pulled into the namespace if a
writer is opened with the "Expected" digest that is already present in the
backend.

The alternative mode, "isolated" requires that clients prove they have access
to the content by providing all of the content to the ingest before the blob is
added to the namespace.

Both modes share backing data, while "shared" will reduce total bandwidth
across namespaces, at the cost of allowing access to any blob just by knowing
its digest.

Signed-off-by: Stephen J Day <stephen.day@docker.com>