server: export unique identifier for instance
The change provides a plugin for the containerd instance to export a
unique id for each containerd instance. The id is such that if the
entire root directory, such as `/var/lib/containerd`, is removed, the
instance will get a new id. This is also a solid place to export other
values that describe the configuration of the containerd instance.

Signed-off-by: Stephen J Day <stephen.day@docker.com>