RELEASES.md: update components for 1.0 release
This should have been updated at release time, but the GRPC API and the
metrics API is considered stable as of the 1.0 release. This means that
changes that break compatibility with previous versions of containerd
will no longer be accepted.

As part of this update, I am also proposing that we push out Go API
stability to 1.2 so that we can incorporate feedback from the Windows
integration in 1.1.

Signed-off-by: Stephen J Day <stephen.day@docker.com>