containerd-shim: link with GNU ld
This change adds `-linkmode=external` option to `go build -ldflags`,
so GNU linker is used instead of Go's own. The change is done so
the resulting binary will have `.note.gnu.build-id` which is now
required by Fedora 27 rpm-build system (and will most probably
be required in RHEL 8).

Unfortunately this results in size increase; here's `size` output
before and after the change:

 5043206	 126020	 165248	5334474	 5165ca	./bin/containerd-shim-new
 4191140	 117552	 139392	4448084	 43df54	./bin/containerd-shim-old

I tried a few options to make the size smaller (like `-Wl,--as-needed`
and a few others) to no avail.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>